refactor: move `merge` to `utils-shared`
