+++ packages/fatcher/src/fatcher.ts
@@ -3,8 +3,7 @@ import { defaultOptions, mergeOptions } from './options';
 import { fetcher, registerMiddlewares, composeMiddlewares } from './middlewares';
 import { createContext } from './context';
 import { canActivate } from './helpers';
-import { merge } from './utils';
-import { isFunction } from '@fatcherjs/utils-shared';
+import { isFunction, merge } from '@fatcherjs/utils-shared';
 
 /**
  * Send HTTP request with custom options.
