+++ packages/fatcher/src/utils.ts
@@ -1,18 +1,4 @@
 /**
- * Merge many objects to one.
- */
-export function merge<T extends Record<string, any>>(
-    initial: T,
-    patches: Partial<T>[],
-    customMerge: (merged: T, patch: Partial<T>) => Partial<T>
-): T {
-    return patches.reduce(
-        (merged, patch) => Object.assign(merged, customMerge(merged, patch)),
-        Object.assign(Object.create(null), initial)
-    );
-}
-
-/**
  * Parse a url with `base` and `url`
  *
  * Supports relative url like `../` and `./`
