+++ packages/utils-shared/src/merge.ts
@@ -1,0 +1,13 @@
+/**
+ * Merge many objects to one.
+ */
+export function merge<T extends Record<string, any>>(
+    initial: T,
+    patches: Partial<T>[],
+    customMerge: (merged: T, patch: Partial<T>) => Partial<T>
+): T {
+    return patches.reduce(
+        (merged, patch) => Object.assign(merged, customMerge(merged, patch)),
+        Object.assign(Object.create(null), initial)
+    );
+}
